pyproject.toml
#
# pyvider-hcl/pyproject.toml
#

[project]
name = "pyvider-hcl"
description = "Pyvider HCL: A wrapper for python-hcl2 integrating with Pyvider's Cty type system."
dynamic = ["version"]
requires-python = ">=3.11"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Tim Perkins", email = "code@tim.life" },
]
maintainers = [
    { name = "provide.io", email = "code@provide.io" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "attrs>=25.3.0",
    "python-hcl2>=7.2.1",
    #"pyvider-cty",
    "pyvider",
    "pyvider-cty",
    "provide-foundation",
    "regex>=2024.11.6",
]

[project.urls]
Homepage = "https://foundry.provide.io/pyvider-hcl/"
Documentation = "https://foundry.provide.io/pyvider-hcl/"
Repository = "https://foundry.provide.io/pyvider-hcl/repo"
Issues = "https://foundry.provide.io/pyvider-hcl/issues"

[dependency-groups]
dev = [
    "provide-testkit[standard,advanced-testing,typecheck,build]",
]
docs = [
    "provide-testkit[docs]",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"} # Specifies that packages are under src

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true # Important for pyvider.hcl as a namespace package

[tool.setuptools.dynamic]
version = {file = "VERSION"}


################################################################################
# Pytest Configuration
################################################################################

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src", "."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
    "unit: marks unit tests",
    "isolated: marks tests that must run in isolation to avoid state pollution",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:cannot collect test class.*:pytest.PytestCollectionWarning",
    "ignore:Module already imported so cannot be rewritten.*:pytest.PytestAssertRewriteWarning",
]
norecursedirs = [
    ".git",
    "build",
    "dist",
    "*.egg-info",
    ".venv",
    "htmlcov",
    ".hypothesis",
]
addopts = "-m 'not isolated' -ra -q --color=yes"

# ---------------------------------------------------------------------------- #
# ruff configuration
# ---------------------------------------------------------------------------- #

[tool.ruff]
line-length = 111
indent-width = 4
target-version = "py311"
extend-exclude = ["mutants", "site"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "ANN", "B", "C90", "SIM", "PTH", "RUF"]
ignore = ["ANN401", "B008", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["pyvider", "tests"]
force-sort-within-sections = true
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ---------------------------------------------------------------------------- #
# mypy configuration
# ---------------------------------------------------------------------------- #

[tool.mypy]
python_version = "3.11"
mypy_path = "src"
strict = true
pretty = true
show_error_codes = true
show_column_numbers = true
warn_unused_ignores = true
warn_unused_configs = true
disallow_untyped_decorators = false
exclude = ["^mutants/"]

[[tool.mypy.overrides]]
module = ["hcl2.*", "pytest.*", "opentelemetry.*"]
ignore_missing_imports = true

# ---------------------------------------------------------------------------- #
# coverage configuration
# ---------------------------------------------------------------------------- #

[tool.coverage.run]
source = ["pyvider.hcl"]
branch = true
parallel = true

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "def __repr__",
    "def __str__",
    "pass",
]

[tool.uv.sources]
provide-foundation = { path = "../provide-foundation", editable = true }
